lesson06[C]: complete
