+++ src/by/it/group310901/baradzin/lesson06/README.md
@@ -2,4 +2,4 @@
 
 - [x] A
 - [x] B
-- [ ] C
+- [x] C
