+++ src/by/it/group310901/baradzin/lesson06/dataC.txt
@@ -1,0 +1,2 @@
+5
+5 3 4 4 2
